Add a try/catch in query.encodeValue
This change enables converting values containing special characters, such as the percent sign (%), to strings, by always encoding values for which the check if encoding is actually necessary has failed.
For values containing URI escape characters, the check throws an error (because they are not decodable), but we still need to encode them.

Fixes persvr/rql#75